requirements.txt
langgraph==1.1.4
langchain-openai==1.1.12
langchain-core==1.2.23
langchain-community==0.4.1
fastapi==0.135.2
uvicorn[standard]==0.42.0
aiohttp==3.13.4
pyyaml==6.0.3
python-dotenv==1.2.2
tavily-python==0.7.23
sqlalchemy==2.0.48
aiosqlite==0.21.0
alembic==1.17.2
redis==7.1.0
pydantic-settings==2.12.0
pytest==9.0.2
pytest-asyncio==1.3.0
httpx==0.28.1
